+++ packages/kit/src/runtime/client/renderer.js
@@ -757,7 +757,7 @@ export class Renderer {
 							`${url.pathname}${url.pathname.endsWith('/') ? '' : '/'}__data.json`,
 							{
 								headers: {
-									'x-sveltekit-noredirect': 'true'
+									'x-sveltekit-load': 'true'
 								}
 							}
 						);