+++ sites/kit.svelte.dev/src/routes/__error.svelte
@@ -14,47 +14,8 @@
 
 	// we don't want to use <svelte:window bind:online> here, because we only care about the online
 	// state when the page first loads
-	let online = typeof navigator !== 'undefined'
-		? navigator.onLine
-		: true;
+	let online = typeof navigator !== 'undefined' ? navigator.onLine : true;
 </script>
-
-<style>
-	.container {
-		padding-inline: var(--side-nav);
-		padding-block: var(--top-offset) 6rem;
-	}
-
-	h1, p {
-		margin-inline: auto;
-		margin-block: 0;
-	}
-
-	h1 {
-		font-size: 2.8em;
-		font-weight: 300;
-		margin: 0;
-		margin-block-end: 0.5em;
-	}
-
-	p {
-		margin-inline: auto;
-		margin-block: 1em;
-	}
-
-	.error {
-		background-color: #da106e;
-		color: white;
-		padding-inline: 16px;
-		padding-block: 12px;
-		font: 600 16px/1.7 var(--font);
-		border-radius: 2px;
-	}
-
-	/* @media (min-width: 480px) {
-		h1 { font-size: 4em }
-	} */
-</style>
 
 <svelte:head>
 	<title>{status}</title>
@@ -77,7 +38,11 @@
 				<p>Please try reloading the page.</p>
 			{/if}
 
-			<p>If the error persists, please drop by <a href="https://svelte.dev/chat">Discord chatroom</a> and let us know, or raise an issue on <a href="https://github.com/sveltejs/svelte">GitHub</a>. Thanks!</p>
+			<p>
+				If the error persists, please drop by <a href="https://svelte.dev/chat">Discord chatroom</a>
+				and let us know, or raise an issue on
+				<a href="https://github.com/sveltejs/svelte">GitHub</a>. Thanks!
+			</p>
 		{/if}
 	{:else}
 		<h1>It looks like you're offline</h1>
@@ -85,3 +50,37 @@
 		<p>Reload the page once you've found the internet.</p>
 	{/if}
 </div>
+
+<style>
+	.container {
+		padding: var(--top-offset) var(--side-nav) 6rem var(--side-nav);
+	}
+
+	h1,
+	p {
+		margin: 0 auto;
+	}
+
+	h1 {
+		font-size: 2.8em;
+		font-weight: 300;
+		margin: 0;
+		margin-bottom: 0.5em;
+	}
+
+	p {
+		margin: 1em auto;
+	}
+
+	.error {
+		background-color: #da106e;
+		color: white;
+		padding: 12px 16px;
+		font: 600 16px/1.7 var(--font);
+		border-radius: 2px;
+	}
+
+	/* @media (min-width: 480px) {
+		h1 { font-size: 4em }
+	} */
+</style>